refactor client notifications utils into jsx
